+++ src/gtirb-decoder/format/ElfLoader.cpp
@@ -133,7 +133,12 @@ void ElfExceptionLoader(const gtirb::Module &Module, DatalogProgram &Program)
 
 ElfExceptionDecoder::ElfExceptionDecoder(const gtirb::Module &module)
 {
-    uint8_t ptrsize(8);
+    uint8_t ptrsize = 8;
+    if(module.getISA() == gtirb::ISA::IA32)
+    {
+        ptrsize = 4;
+    }
+
     std::string ehFrame, ehFrameHeader, gccExcept;
     uint64_t addressEhFrame(0), addressEhFrameHeader(0), addressGccExcept(0);